feat: add support for complex type in $orderBy
