+++ .changeset/purple-ravens-refuse.md
@@ -1,0 +1,5 @@
+---
+'@sap-ux/fe-mockserver-core': patch
+---
+
+Add support for ComplexType in $orderBy
