+++ packages/fe-mockserver-core/src/data/entitySets/entitySet.ts
@@ -9,18 +9,7 @@ import type { Action, EntitySet, EntityType, Property } from '@sap-ux/vocabulari
 import { FunctionBasedMockData } from '../../mockdata/functionBasedMockData';
 import { FileBasedMockData } from '../../mockdata/fileBasedMockData';
 import type { FilterMethodCall, LambdaExpression } from '../../request/filterParser';
-
-function getData(fullData: any, objectPath: string): any {
-    if (fullData === undefined || objectPath.length === 0) {
-        return fullData;
-    }
-    if (objectPath.indexOf('/') === -1) {
-        return fullData?.[objectPath]; // fullData can be null
-    } else {
-        const subObjectPath = objectPath.split('/');
-        return getData(fullData[subObjectPath[0]], subObjectPath.slice(1).join('/'));
-    }
-}
+import { getData } from '../common';
 
 type PreparedFunction = {
     fn: Function;
