+++ packages/fe-mockserver-core/test/unit/v4/services/complexType/A.json
@@ -8,5 +8,12 @@
                 "B": "Hey"
             }
         }
+    },
+    {
+        "ID": "A2",
+        "complex": {
+            "value1": 2,
+            "value2": 1
+        }
     }
 ]
